+++ src/pages/instant-observability.js
@@ -7,7 +7,6 @@ import { css } from '@emotion/react';
 import SegmentedControl from '../components/SegmentedControl';
 import Overlay from '../components/Overlay';
 import PackTile from '../components/PackTile';
-import IOBanner from '../components/IOBanner';
 import IOLogo from '../components/IOLogo';
 import QuickstartFilter from '../components/quickstarts/QuickstartFilter';
 import {
@@ -23,6 +22,7 @@ import BUILD_YOUR_OWN from '../images/build-your-own.svg';
 import GUIDED_INSTALL from '../images/guided-install.svg';
 import { useDebounce } from 'react-use';
 import { sortFeaturedQuickstarts } from '../utils/sortFeaturedQuickstarts';
+import SuperTilesExperiment from '../experiments/super_tiles';
 
 import {
   QUICKSTARTS_REPO,
@@ -294,7 +294,7 @@ const QuickstartsPage = ({ data, location }) => {
             }
           `}
         >
-          {isMobile && <IOBanner isMobile />}
+          {isMobile && <SuperTilesExperiment isMobile />}
           <div
             css={css`
               padding: var(--site-content-padding);
@@ -410,7 +410,7 @@ const QuickstartsPage = ({ data, location }) => {
             padding: var(--site-content-padding);
           `}
         >
-          {!isMobile && <IOBanner />}
+          {!isMobile && <SuperTilesExperiment />}
           <div
             css={css`
               background-color: var(--secondary-background-color);